+++ branch-beta/CHANGELOG.md
@@ -7,6 +7,7 @@
 - jslint - try to improve parser to be able to parse jquery.js without stopping.
 
 # v2025.12.1-beta
+- jslint - Allow parenthesis after ellipsis inside a function call.
 - ci - Upgrade nodejs used in ci to v24.
 
 # v2025.10.31